--- scripts/maintenance/computational-clusters/requirements.txt
@@ -1,17 +1,0 @@
-arrow
-asyncpg
-black
-boto3
-# NOTE: these must be in sync with osparc
-cloudpickle==2.2.1
-dask[distributed]==2023.3.2
-mypy_boto3_ec2
-types-boto3
-parse
-paramiko
-pydantic[email]
-pylint
-python-dotenv
-typer[all]
-sqlalchemy[asyncio]
-sshtunnel
